+++ CMakeLists.txt
@@ -10,7 +10,7 @@ set(CMAKE_SYSTEM_NAME Generic)
 set(CMAKE_C_COMPILER_WORKS 1)
 set(CMAKE_CXX_COMPILER_WORKS 1)
 
-project("synth" C CXX)
+project("teensy-toy-synth" C CXX)
 
 include(CMakeListsPrivate.txt)
 
@@ -30,7 +30,4 @@ add_custom_target(
     WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
 )
 
-add_executable(Z_DUMMY_TARGET ${SRC_LIST} include/luts.h include/io/MCP23008.h src/lcd16x2.cpp src/I2C1602.h
-        src/I2C1602.cpp src/SerialLCDWriter.cpp include/SerialLCDWriter.h src/Controller.cpp include/io/Controller.h
-        include/LinkedList.h include/io/InputControllers.h include/synth_globals.h src/synthesizer.cpp
-        include/synthesizer/synthesizer.h include/synthesizer/components.h src/InputSnapshot.cpp include/io/InputSnapshot.h src/synth_globals.cpp src/VirtualInput.cpp include/io/VirtualInput.h src/InputListener.cpp include/io/InputListener.h src/InputPollster.cpp include/io/InputPollster.h include/io/toyIO.h include/synth_defines.h)
+add_executable(Z_DUMMY_TARGET ${SRC_LIST})
