+++ CMakeListsPrivate.txt
@@ -18,7 +18,7 @@ SET(CMAKE_C_FLAGS "-Wall -ffunction-sections -fdata-sections -mthumb -mcpu=corte
 set(CMAKE_CXX_STANDARD 14)
 
 if (CMAKE_BUILD_TYPE MATCHES "teensy41")
-    add_definitions(-D'PLATFORMIO=50101')
+    add_definitions(-D'PLATFORMIO=50200')
     add_definitions(-D'__IMXRT1062__')
     add_definitions(-D'ARDUINO_TEENSY41')
     add_definitions(-D'USB_SERIAL')
@@ -37,10 +37,10 @@ if (CMAKE_BUILD_TYPE MATCHES "teensy41")
     include_directories("$ENV{HOME}/.platformio/packages/framework-arduinoteensy/libraries/SdFat/src")
     include_directories("$ENV{HOME}/.platformio/packages/framework-arduinoteensy/libraries/SerialFlash")
     include_directories("$ENV{HOME}/.platformio/packages/framework-arduinoteensy/libraries/SPI")
-    include_directories("${CMAKE_CURRENT_LIST_DIR}/.pio/libdeps/teensy41/LiquidCrystal_I2C")
     include_directories("$ENV{HOME}/.platformio/packages/framework-arduinoteensy/libraries/Wire")
     include_directories("$ENV{HOME}/.platformio/packages/framework-arduinoteensy/libraries/Wire/utility")
     include_directories("$ENV{HOME}/.platformio/packages/framework-arduinoteensy/cores/teensy4")
+    include_directories("${CMAKE_CURRENT_LIST_DIR}/.pio/libdeps/teensy41/LiquidCrystal_I2C")
     include_directories("$ENV{HOME}/.platformio/packages/framework-arduinoteensy/libraries/ADC")
     include_directories("$ENV{HOME}/.platformio/packages/framework-arduinoteensy/libraries/AccelStepper/src")
     include_directories("$ENV{HOME}/.platformio/packages/framework-arduinoteensy/libraries/Adafruit_NeoPixel")
