+++ include/synthesizer/synthesizer.h
@@ -9,6 +9,7 @@
 
 
 
+
 void synth_init();
 
 #endif //SYNTH_SYNTHESIZER_H
